refactor(resource form): move form components into dir
