+++ src/pages/resource/add-edit-resource/ResourceForm.js
@@ -3,17 +3,17 @@ import DialogContent from '@material-ui/core/DialogContent'
 import PropTypes from 'prop-types'
 import React, { Fragment } from 'react'
 
-import { Actions } from './Actions'
-import { Attachment } from './Attachment'
-import { File } from './File'
 import { FormDialog } from '../../../components/form/FormDialog'
 import { FormSection } from '../../../components/form/FormSection'
-import { Name } from './Name'
-import { TitleDetails } from './TitleDetails'
-import { TitleResource } from './TitleResource'
-import { Type } from './Type'
-import { Url } from './Url'
 import { resourceTypes } from '../../../utils/resource/constants'
+import { Actions } from './ResourceForm/Actions'
+import { Attachment } from './ResourceForm/Attachment'
+import { File } from './ResourceForm/File'
+import { Name } from './ResourceForm/Name'
+import { TitleDetails } from './ResourceForm/TitleDetails'
+import { TitleResource } from './ResourceForm/TitleResource'
+import { Type } from './ResourceForm/Type'
+import { Url } from './ResourceForm/Url'
 
 export const ResourceForm = props => (
     <FormDialog title={props.title} open={props.open} onClose={props.onCancel}>
