+++ src/pages/resource/add-edit-resource/ResourceForm/Attachment.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import i18n from '@dhis2/d2-i18n'
 
-import { FormRow } from '../../../components/form/FormRow'
-import { RadioButtons } from '../../../components/form/RadioButtons'
+import { FormRow } from '../../../../components/form/FormRow'
+import { RadioButtons } from '../../../../components/form/RadioButtons'
 
 export const Attachment = () => (
     <FormRow>
