+++ src/pages/resource/add-edit-resource/ResourceForm/File.js
@@ -2,9 +2,9 @@ import PropTypes from 'prop-types'
 import React from 'react'
 import i18n from '@dhis2/d2-i18n'
 
-import { File as FileInput } from '../../../components/form/File'
-import { FormRow } from '../../../components/form/FormRow'
-import { isRequired } from '../../../utils/form/validators'
+import { File as FileInput } from '../../../../components/form/File'
+import { FormRow } from '../../../../components/form/FormRow'
+import { isRequired } from '../../../../utils/form/validators'
 
 export const File = props => (
     <FormRow>
