+++ src/pages/resource/add-edit-resource/ResourceForm/Title.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { FormSectionTitle } from '../../../components/form/FormSectionTitle'
+import { FormSectionTitle } from '../../../../components/form/FormSectionTitle'
 
 export const Title = props => <FormSectionTitle>{props.label}</FormSectionTitle>
