+++ src/pages/resource/add-edit-resource/ResourceForm/TitleDetails.js
